simd accelerated tokenization
