+++ models/tokenator/Cargo.toml
@@ -4,12 +4,12 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-serde = { version = "1.0.205", features = ["derive"] }
-serde_json = "1.0.122"
-colored = "2.1.0"
-serde-big-array = "0.5.1"
-regex = { version = "1.10.6", features = ["perf-dfa-full"] }
-tracing = "0.1.40"
+# serde = { version = "1.0.205", features = ["derive"] }
+# serde_json = "1.0.122"
+# colored = "2.1.0"
+# serde-big-array = "0.5.1"
+# regex = { version = "1.10.6", features = ["perf-dfa-full"] }
+# tracing = "0.1.40"
 
 
 [features]
@@ -17,14 +17,14 @@ never-inline = []
 skip-list = []
 
 [dev-dependencies]
-criterion = "0.3"
-rand = "0.8.5"
-tokenizers = "0.20.0"
-pretty_assertions = "1.4.1"
-postcard = { version = "1.0.10", features = ["use-std"] }
-reqwest = { version = "0.12.7", features = ["blocking"] }
-tracing-subscriber = "0.3.18"
+# criterion = "0.3"
+# rand = "0.8.5"
+# tokenizers = "0.20.0"
+# pretty_assertions = "1.4.1"
+# postcard = { version = "1.0.10", features = ["use-std"] }
+# reqwest = { version = "0.12.7", features = ["blocking"] }
+# tracing-subscriber = "0.3.18"
 
-[[bench]]
-name = "tokenize"
-harness = false
+# [[bench]]
+# name = "tokenize"
+# harness = false
